Better wpChunk.push type, move webpackChunk types
