+++ src/types/webpack.ts
@@ -1,4 +1,52 @@
-import type { ModuleExports, RawModule } from "./discord";
+export type ObjectExports = Record<string, unknown>;
+
+export type ModuleExports =
+  | ObjectExports
+  | ((...args: unknown[]) => unknown)
+  | string
+  | boolean
+  | symbol;
+export type ModuleExportsWithProps<P extends string> = Record<P, unknown> &
+  Record<PropertyKey, unknown>;
+
+export interface RawModule {
+  id: number;
+  loaded: boolean;
+  exports: ModuleExports;
+}
+
+export interface RawModuleWithProps<P extends string> extends RawModule {
+  exports: ModuleExportsWithProps<P>;
+}
+
+export type WebpackRequireCache = Record<string | number, RawModule>;
+
+export type WebpackRequire = ((e: number) => ModuleExports) & {
+  c: WebpackRequireCache;
+  d: (module: ModuleExports, exports: Record<string, () => unknown>) => void;
+  m: WebpackChunk;
+};
+
+export type WebpackModule = (
+  wpModule: RawModule,
+  wpExports: typeof wpModule.exports,
+  wpRequire: WebpackRequire,
+) => void;
+
+export type WebpackChunk<T extends ((r: WebpackRequire) => unknown) | undefined = undefined> = [
+  Array<symbol | number>,
+  Record<number, WebpackModule>,
+  T?,
+];
+
+// Do NOT put `WebpackChunk[]` first, otherwise TS
+// prioritizes Array.prototype.push over this custom
+// push method and starts producing errors.
+export type WebpackChunkGlobal = {
+  push<T extends ((r: WebpackRequire) => unknown) | undefined = undefined>(
+    chunk: WebpackChunk<T>,
+  ): T extends undefined ? undefined : ReturnType<NonNullable<typeof chunk[2]>>;
+} & WebpackChunk[];
 
 export type Filter = (module: RawModule) => boolean | ModuleExports;
